Compression benchmarks and tests refactor.
