+++ CMakeLists.txt
@@ -62,36 +62,51 @@ target_link_libraries(LibEugeneCore
         CONAN_PKG::catch2)
 
 if (DEFINED ENV{EUGENE_BUILD_TESTS})
-    macro(eugene_test component path_prefix)
+	message("-- Building core unit tests")
+
+
+	macro(eugene_test component path_prefix)
         add_executable(Test${component} ${path_prefix}/${component}.cpp)
 		    target_link_libraries(Test${component} TestMain LibEugeneCore)
         add_test(NAME Test${component} COMMAND Test${component})
         message("   -- Test about '${component}' component added from \"${path_prefix}/${component}.cpp\"")
     endmacro()
 
-    message("-- Building core unit tests")
-
     add_library(TestMain src/core/Main.cpp)
     target_link_libraries(TestMain CONAN_PKG::catch2)
 
-    eugene_test(Compression src/core/storage/compression)
     eugene_test(Btree src/core/storage/btree)
-	  eugene_test(Node src/core/storage/btree)
+	eugene_test(Node src/core/storage/btree)
     eugene_test(Page src/core/storage)
     eugene_test(PageCache src/core/storage)
+
+
+	add_library(LibTestsShared src/core/storage/compression/tests/Shared.h)
+	set_target_properties(LibTestsShared PROPERTIES LINKER_LANGUAGE CXX)
+	target_include_directories(LibTestsShared PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/src/core/storage/compression/tests")
+
+	eugene_test(CompDecompTests src/core/storage/compression/tests)
+	eugene_test(CompTests src/core/storage/compression/tests)
+	eugene_test(DecompTests src/core/storage/compression/tests)
 endif()
 
 if (DEFINED ENV{EUGENE_BUILD_BENCHMARKS})
-    add_executable(BtreeBench src/core/storage/btree/Bench.cpp)
+	message("-- Building benchmarks")
+
+
+	add_executable(BtreeBench src/core/storage/btree/Bench.cpp)
     target_link_libraries(BtreeBench LibEugeneCore)
 
-    add_executable(bench_comp src/core/storage/compression/BenchCompressor.cpp)
-    target_link_libraries(bench_comp PUBLIC LibEugeneCore)
 
-	add_executable(bench_decomp src/core/storage/compression/BenchDecompressor.cpp)
-	target_link_libraries(bench_decomp PUBLIC LibEugeneCore)
+	add_library(LibCompShared src/core/storage/compression/benchmarks/Shared.h)
+	set_target_properties(LibCompShared PROPERTIES LINKER_LANGUAGE CXX)
+	target_include_directories(LibCompShared PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/src/core/storage/compression/benchmarks")
 
-    message("-- Building benchmarks")
+    add_executable(bench_comp src/core/storage/compression/benchmarks/CompBench.cpp)
+    target_link_libraries(bench_comp PUBLIC LibEugeneCore LibCompShared)
+
+	add_executable(bench_decomp src/core/storage/compression/benchmarks/DecompBench.cpp)
+	target_link_libraries(bench_decomp PUBLIC LibEugeneCore LibCompShared)
 endif()
 
 add_executable(main src/standalone/json-parse/main.cpp)
