+++ src/core/storage/compression/tests/CompDecompTests.cpp
@@ -1,0 +1,38 @@
+#include <core/storage/compression/tests/Shared.h>
+
+void folder_test(std::map<std::string, std::string> &params, const int text_size) {
+	for (const auto &item : params)
+		REQUIRE(!exists(item.second));
+
+	REQUIRE(create_testing_directory(params["test_dir_name"], text_size));
+	REQUIRE(exists(params["test_dir_name"]));
+
+	compression::Compressor compress{
+		std::vector<std::string>(1, params["test_dir_name"]),
+		params["compressed_name"]
+	};
+	compress();
+	REQUIRE(exists(params["compressed_name"]));
+
+	fs::rename(params["test_dir_name"], params["changed_to_initial_dir"]);
+	REQUIRE(exists(params["changed_to_initial_dir"]));
+
+	decompression::Decompressor decompress{params["compressed_name"]};
+	decompress();
+	REQUIRE(exists(params["test_dir_name"]));
+
+	REQUIRE(compare_folders("InitialDir", "ForTesting"));
+
+	check_initial_compressed_size(params["test_dir_name"], params["compressed_name"]);
+	REQUIRE(clean(params));
+}
+
+TEST_CASE("CompDecomp comp_decomp", "[compressor_decompressor]") {
+	std::map<std::string, std::string> params;
+	params["test_dir_name"] = "ForTesting";
+	params["changed_to_initial_dir"] = "InitialDir";
+	params["compressed_name"] = "Test";
+
+	for (int i = 1; i < 5; ++i)
+		folder_test(params, (int) pow(10, i));
+}
